requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
paddlepaddle
paddleocr
numpy
